Reparando Destroy Products

+++ src/controllers/productController.js
@@ -65,7 +65,6 @@ module.exports = {
     // VISTA QUE MUESTRA LOS PRODUCTOS DISPONIBLES EN MODELOS PARA IMPRIMIR --OK
     reqPersonalizados: (req, res) => {
         
-        console.log("  --------------que trae params------------->>> " + req.params)
         db.Product.findAll({
             include: [
                 {
@@ -122,8 +121,6 @@ module.exports = {
             }
         } 
 
-
-
       if(Object.entries(errors).length === 0){
            const {name, price, discount, heigth, time, categoryId, materialId,description,imagen,view} = req.body;
 
@@ -161,13 +158,13 @@ module.exports = {
         .catch(error => console.log("======ERROR========>" + error))
        }else{
 
-/*             if (req.files.length > 0){
+            if (req.files.length > 0){
                 req.files.forEach(({filename}) => {
                     fs.existsSync(path.resolve(__dirname,"..","..","public","images","imgProducts",filename)) &&
                     fs.unlinkSync(path.resolve(__dirname,"..","..","public","images","imgProducts",filename))
                 });
             } 
- */
+
             let categories = db.Category.findAll()
             let materials = db.Material.findAll()
             Promise.all([categories,materials])
@@ -184,9 +181,6 @@ module.exports = {
 
 
 
-
-
-    
     detalle: (req, res) => {
 
         db.Product.findByPk(req.params.id,{
